requirements.txt
#Libraries to use with docker
pandas
python-dotenv
fastapi
uvicorn
scrapy
beautifulsoup4
python-multipart
openpyxl
xlsxwriter